Refactored bootstrap file
