+++ backend/product.rb
@@ -1,0 +1,8 @@
+class Product
+  attr_reader :name, :type
+
+  def initialize(name:, type:)
+    @name = name
+    @type = type
+  end
+end
