+++ backend/product_order/physical.rb
@@ -1,0 +1,10 @@
+require_relative 'item'
+require_relative '../shipping_label'
+
+module ProductOrder
+  class Physical < Item
+    def deliver(order)
+      ShippingLabel.new(product: self, address: order.address).generate
+    end
+  end
+end
